+++ src/JIM.PostgresData/Migrations/20260624105858_AddExpressionToExampleDataTemplateAttribute.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace JIM.PostgresData.Migrations
+{
+    /// <inheritdoc />
+    public partial class AddExpressionToExampleDataTemplateAttribute : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "Expression",
+                table: "ExampleDataTemplateAttributes",
+                type: "text",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Expression",
+                table: "ExampleDataTemplateAttributes");
+        }
+    }
+}